# Make small retrieval 200x faster