--- a/markets/dagstore/piecereader.go
+++ b/markets/dagstore/piecereader.go
@@ -110,6 +110,8 @@ func (p *pieceReader) ReadAt(b []byte, off int64) (n int, err error) {
 			p.r = nil
 		}
 
+		log.Debugw("pieceReader new stream", "at", p.rAt, "off", off-p.rAt)
+
 		p.rAt = off
 		p.r, _, err = p.api.FetchUnsealedPiece(p.ctx, p.pieceCid, uint64(p.rAt))
 		if err != nil {