@raulk 2021-11-26
@magik6k Oh yeah, that's a good point. Both the layout of data in disk and the retrieval order are depth-first. That fact, along with the "burn" leniency, should solve both challenges in majority of cases 👍  Maybe we can think of some selectors and DAG combinations whose pattern of access would be different, and test those?